Save database as a json format
Previously, the format of database file was jl (JSON Lines) to make
append new entry easier.
But now the saving method is overwriting entire database rather than
appending new entry to prevent the database grows larger and larger.
So the database format is changed to json format  because there is
no need to keep the jl format.